pharmaconnect/requirements-dev.txt
# Dépendances de développement et tests
-r requirements.txt

# Tests
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.6.1

# Qualité de code
black==24.8.0
flake8==7.1.1
isort==5.13.2

# Debug
django-debug-toolbar==4.4.6

# Outils développement
ipython==8.27.0
jupyter==1.0.0